# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Zenith Alpha Capital API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/me/kyc:
    post:
      operationId: submitKyc
      tags: [users]
      summary: Submit KYC verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycInput"
      responses:
        "200":
          description: KYC submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── Investment Plans ────────────────────────────────────────────────────────
...
  /portfolios:
    get:
      operationId: listPortfolios
      tags: [investments]
      summary: List user's active investments
      responses:
        "200":
          description: List of portfolios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Portfolio"
    post:
      operationId: createPortfolio
      tags: [investments]
      summary: Invest in a plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PortfolioInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"
  /portfolios/{id}:
    get:
      operationId: getPortfolio
      tags: [investments]
      summary: Get portfolio details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Portfolio"

  # ── Transactions ────────────────────────────────────────────────────────────
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List user transactions
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [deposit, withdrawal, profit, bonus, referral]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, completed]
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Request deposit or withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get transaction details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  # ── Referrals ───────────────────────────────────────────────────────────────
...
  /admin/transactions:
    get:
      operationId: listAllTransactions
      tags: [admin]
      summary: List all transactions (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, completed]
      responses:
        "200":
          description: All transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, email, role, status, balance, totalInvested, totalProfit, referralCode, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, suspended, pending]
        balance:
          type: number
        totalInvested:
          type: number
        totalProfit:
          type: number
        referralCode:
          type: string
        referredBy:
          type: ["string", "null"]
        kycStatus:
          type: string
          enum: [none, pending, approved, rejected]
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        country:
          type: string
    KycInput:
      type: object
      required: [fullName, dateOfBirth, documentType, documentNumber]
      properties:
        fullName:
          type: string
        dateOfBirth:
          type: string
        documentType:
          type: string
          enum: [passport, drivers_license, national_id]
        documentNumber:
          type: string
        address:
          type: string
...
    Portfolio:
      type: object
      required: [id, userId, planId, planName, amount, currentValue, profit, startDate, endDate, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        planId:
          type: integer
        planName:
          type: string
        amount:
          type: number
        currentValue:
          type: number
        profit:
          type: number
        returnRate:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [active, completed, cancelled]
        createdAt:
          type: string
    PortfolioInput:
      type: object
      required: [planId, amount]
      properties:
        planId:
          type: integer
        amount:
          type: number
    Transaction:
      type: object
      required: [id, userId, type, amount, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [deposit, withdrawal, profit, bonus, referral]
        amount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected, completed]
        description:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
        walletAddress:
          type: ["string", "null"]
        rejectReason:
          type: ["string", "null"]
        createdAt:
          type: string
    TransactionInput:
      type: object
      required: [type, amount]
      properties:
        type:
          type: string
          enum: [deposit, withdrawal]
        amount:
          type: number
        paymentMethod:
          type: string
        walletAddress:
          type: string
        description:
          type: string
...