# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zenith Alpha Capital API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  firstName?: string;
  lastName?: string;
  phone?: string;
  country?: string;
}